+++ build_env.py
@@ -9,23 +9,20 @@ try:
             "install",
             "-y",
             "docker-compose",
-            "python3.11-venv",
+            "python3-venv",
+            "python3-pip",
             "acl",
         ],
         check=True,
     )
 
-    subprocess.run(["python3", "-m", "venv", "nostpyenv"], check=True)
-
-    # Activate the virtual environment and run subsequent commands within it (runs prepare server script in venv)
-    activate_cmd = ". nostpyenv/bin/activate && "
     commands = [
         "pip install --upgrade pip",
         "pip install -r requirements.txt",
-        "python prepare_server.py",
+        "python3 prepare_server.py",
     ]
     for cmd in commands:
-        subprocess.run(["sudo", "bash", "-c", activate_cmd + cmd], check=True)
+        subprocess.run(["sudo", "bash", "-c", cmd], check=True)
 
 except subprocess.CalledProcessError as e:
     print(f"An error occurred while executing the command: {e}")